refactor: group classes with package
