+++ src/main/java/vendingmachine/Application.java
@@ -1,6 +1,6 @@
 package vendingmachine;
 
-import java.util.List;
+import vendingmachine.domain.VendingMachine;
 
 public class Application {
     public static void main(String[] args) {
