+++ src/main/java/vendingmachine/domain/CoinCollection.java
@@ -1,4 +1,6 @@
-package vendingmachine;
+package vendingmachine.domain;
+
+import vendingmachine.domain.Coin;
 
 import java.util.ArrayList;
 import java.util.List;
